requirements.txt
asgiref==3.6.0
dj-database-url==0.5.0
Django==3.2.19
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
stripe==5.4.0
